Refactor of getOffsetFromRVA to return errors
Returning -1 as an error does not work in all cases.  Some call sites
may need to be updated to allow only slightly corrupt files through, but
those fields may need an indicator that they are valid or not.

Changed the the name of parseHeader to readOffset to better reflect what
it actually does, which is to just binary.Read() bytes into an existing
struct.  Added readRVA() as a convenience to hopefully reduce errors by
making the difference between file offset and RVA more explicit.